Adding sample code ShopForce
Two objects: List(a shopping list) and Items (shopping items on the
list). When you select an item (e.g.:soap) the trigger adds related
items to the list (custom setting data). A custom button 'Add Item'
when clicked displays VF page with list of items and checkboxes. When
you select or unselect them and hit save you can see the updates on the
Item related list on the List object.